Optimise day 1 part 2
There wasn't any need to calculate the sum within the window. Since the two
windows overlapped, we only need to compare the sums of the non-overlapping
values which were at idx and idx+3.

Map those into an array of true/false to indicate whether or not there was an
increase and count the number of true values.